run-tests: stop explicit expansion of time data
We are about to record more complex time-related data, which will require
repeated changes of every loop touching times data. That will also extend such
lines to a point where things become too long. Instead, we iterate on each entry
and expand values in the loops. We keep intermediate variables in most cases to
preserve readability.

The loop producing json data does a strange inversion for no obvious reason. We
preserved that for now and will fix it in another changeset.