run-tests: use items() instead of iteritems()
This fixes the python3.5 build. We'll presumably want to build our own
helper function or use 2to3 for this on the main source tree, but for
run-tests we want a single-source version that works in 2.6 and 3.5.

--HG--
extra : rebase_source : b06fa228d1aed1b21033a31b13d4f48fd8a52352