record: add an operation arguments to customize recording ui
This patch is part of a series of patches to change the recording ui to reflect
the operation currently running (commit, shelve, revert ...).
This patch adds a new argument to the recording function to reflect in the UI
what operation we are running.

--HG--
extra : histedit_source : e4b2c58912aee2a8413ab900c5ec20a04b08c965