merge: move merge step to the end
Resolving other conflicts before merge ones is better because the state before
the merge is as consistent as possible. It will also help with future work
involving automatic resolution of merge conflicts with an external merge
driver.

There are no ordering issues here because it is easy to verify that the same
file is never in both the dg/dm and the m sets.

--HG--
extra : rebase_source : b01e117ceabd6ab0196764a662e1f7597fdb33a8