filectx: allow custom comparators
We're going to introduce other sorts of filectxes very soon, and we'd like the
cmp method to function properly (i.e. commutatively) for them. The only way to
make that happen is for this cmp method to call into that specialized one if
that defines a custom comparator.

--HG--
extra : rebase_source : 3770e5ada5541a86c41a526e76d599aef3ee959a