mergestate._resolve: handle change/delete conflicts
We will represent a deleted file as 'nullhex' in the in-memory and on-disk
merge states. We need to be able to create absentfilectxes in that case, and
delete the file from disk rather than try to write it out.

--HG--
extra : rebase_source : f49d9c7f899e39b64efc023863ff779aab34d762