ui: add method to return option and all sub-options
Apparently ":" has been blessed as a generic separator for
options and sub-options. We formalize this by introducing an API
for obtaining an option and all its sub-options.

This will be used in a subsequent patch for declaring sub-options
for [paths].

--HG--
extra : rebase_source : 05b3118732f2a260690009cbc3b38d325d6cb7bf