verify: move verify logic into a class
In order to allow extensions to hook into the verification logic more easily, we
need to refactor it into multiple functions. The first step is to move it to a
class so the shared state can be more easily accessed.

--HG--
extra : rebase_source : db548fc69ab1383330c37f14dc4036bab21c2274