verify: clean up weird error/warning lists
Nested functions in Python are not able to assign to variables in the
outer scope without something like the list trick because assignments
refer to the inner scope. So, we formerly used a list to give an
object to assign into.

Now that error and warning are object members, the [0] hack is no
longer needed.

--HG--
extra : amend_source : 5cb39f3cec9e9715fc77f0dd434d883c7f2b16a8