tests: non-GNU mv can't move a file onto a hardlink of itself
On Solaris, when you try to mv a file to another path that is a hardlink to
the original, it complains that they're identical.  GNU mv doesn't
complain, but it's simpler to just remove the original file instead.

--HG--
extra : rebase_source : dd1d4da40143d846e4a7d8262f18dd444d2421e7